output-processor/requirements.txt
confluent-kafka
jq
psycopg2-binary
prometheus-client